backend/requirements.txt
# SF Parking Data Pipeline Dependencies
requests>=2.31.0
aiohttp>=3.9.0
pandas>=2.1.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
pydantic>=2.5.0
schedule>=1.2.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
